+++ library/feature/support/src/main/kotlin/com/mihaicristiancondrea/android/libs/apptoolkit/feature/support/data/mappers/ProductDetailsExtensions.kt
@@ -15,7 +15,7 @@
  * along with this program. If not, see <https://www.gnu.org/licenses/>.
  */
 
-package com.mihaicristiancondrea.android.libs.apptoolkit.feature.support.utils.extensions
+package com.mihaicristiancondrea.android.libs.apptoolkit.feature.support.data.mappers
 
 import com.android.billingclient.api.BillingClient
 import com.android.billingclient.api.ProductDetails