+++ sample/feature/onboarding/src/main/kotlin/com/mihaicristiancondrea/android/apps/apptoolkit/feature/onboarding/ui/AppOnboardingProvider.kt
@@ -24,6 +24,7 @@ import androidx.compose.material.icons.outlined.AccountCircle
 import androidx.compose.material.icons.outlined.Build
 import androidx.compose.material.icons.outlined.Star
 import com.mihaicristiancondrea.android.apps.apptoolkit.feature.onboarding.R
+import com.mihaicristiancondrea.android.apps.apptoolkit.feature.onboarding.domain.models.OnboardingKeys
 import com.mihaicristiancondrea.android.libs.apptoolkit.feature.onboarding.ui.models.OnboardingPage
 import com.mihaicristiancondrea.android.libs.apptoolkit.feature.onboarding.ui.views.pages.finish.FinishOnboardingPage
 import com.mihaicristiancondrea.android.libs.apptoolkit.feature.onboarding.ui.views.pages.firebase.FirebaseOnboardingPage